Added nifty Hotkey GUI.
Displays what items are currently set to the hotkey slots.

If the item stacks (e.g. potions), it displays the number of this item
in the inventory.

If the item is equipment (e.g. sword), it displays whether or not the
item is equipped.

Very messy code, will polish up eventually.