Reworked Hotkeys
The hotkey system works in the same way, but the code is a bit cleaner.
Added a new class for hotkeys and reworked the system using it.
Considering redoing the inventory system with a similar way.

Hotkeys are 1, 2, 3, and 4.
Press 5 or an unassigned hotkey to configure.
Press H to view the current hotkeys.